Modified type converter and moved all converters to helper package
